requirements.txt
anthropic
boto3
fastapi
httpx
PyJWT[crypto]
cryptography
faster-whisper
matplotlib==3.9.4
mpmath
notion-client
matplotlib==3.9.4
numpy==1.26.4
paho-mqtt
openai==2.3.0
paho-mqtt
pydantic
pyarrow==16.1.0
pre-commit==4.3.0
psutil
openai==1.30.1
pandas
pydantic
pyarrow==16.1.0
pre-commit==4.3.0
PyJWT
PyYAML==6.0.2
python-dotenv
qiskit-aer==0.17.1
qiskit-machine-learning==0.8.4
requests==2.32.4
redis
scipy
streamlit==1.48.1
torch
uvicorn
whisper
typer
pytest
openai-whisper
torch
matplotlib==3.8.4
numpy==1.26.4
openai==1.30.1
pyyaml==6.0.1
requests==2.32.2
streamlit==1.35.0
torch
whisper
typer==0.12.3
pytest
